+++ tests/functions/tests.primitivesvalues.prod/PrimitivesValuesProdFunc.cpp
@@ -199,7 +199,7 @@ bool PrimitivesValuesProdFunction::runStep(const openfluid::base::SimulationStat
 
   TUID = TU->getID();
 
-  TheDouble = TUID/10;
+  TheDouble = (double)TUID/10;
   TheLong = TUID;
   TheBool = (TUID%2 == 0);
   TheString = Glib::ustring::compose("ID %1",TUID);
